requirements.txt
# Streaming Instagram Captioner Requirements

# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=9.0.0

# Optional dependencies for enhanced functionality
scikit-learn>=1.0.0
sentence-transformers>=2.2.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Configuration
PyYAML>=6.0.0

# Logging and monitoring
tqdm>=4.65.0


